+++ .golangci.yml
@@ -22,7 +22,7 @@ linters:
     - misspell
     - staticcheck
     - structcheck
-#    - typecheck
+    - typecheck
     - unused
     - unconvert
     - varcheck
